spread code out a bit more; improved docs
