+++ worker/fortress/fortress_test.go
@@ -7,14 +7,12 @@ import (
 	"sync"
 	"time"
 
-	"github.com/juju/errors"
 	"github.com/juju/testing"
 	jc "github.com/juju/testing/checkers"
 	gc "gopkg.in/check.v1"
 
 	coretesting "github.com/juju/juju/testing"
 	"github.com/juju/juju/worker"
-	"github.com/juju/juju/worker/dependency"
 	"github.com/juju/juju/worker/fortress"
 )
 
@@ -24,14 +22,25 @@ type FortressSuite struct {
 
 var _ = gc.Suite(&FortressSuite{})
 
+func (s *FortressSuite) TestOutputBadSource(c *gc.C) {
+	fix := newFixture(c)
+	defer fix.TearDown(c)
+
+	var dummy struct{ worker.Worker }
+	var out fortress.Guard
+	err := fix.manifold.Output(dummy, &out)
+	c.Check(err, gc.ErrorMatches, "in should be \\*fortress\\.fortress; is .*")
+	c.Check(out, gc.IsNil)
+}
+
 func (s *FortressSuite) TestOutputBadTarget(c *gc.C) {
 	fix := newFixture(c)
 	defer fix.TearDown(c)
 
-	var state interface{}
-	err := fix.manifold.Output(fix.worker, &state)
+	var out interface{}
+	err := fix.manifold.Output(fix.worker, &out)
 	c.Check(err.Error(), gc.Equals, "out should be *fortress.Guest or *fortress.Guard; is *interface {}")
-	c.Check(state, gc.IsNil)
+	c.Check(out, gc.IsNil)
 }
 
 func (s *FortressSuite) TestStoppedUnlock(c *gc.C) {
@@ -334,81 +343,4 @@ func (s *FortressSuite) TestAbortedLockdownUnlock(c *gc.C) {
 	err = guard.Unlock()
 	c.Assert(err, jc.ErrorIsNil)
 	AssertUnlocked(c, fix.Guest(c))
-}
-
-type fixture struct {
-	manifold dependency.Manifold
-	worker   worker.Worker
-}
-
-func newFixture(c *gc.C) *fixture {
-	manifold := fortress.Manifold()
-	worker, err := manifold.Start(nil)
-	c.Assert(err, jc.ErrorIsNil)
-	return &fixture{
-		manifold: manifold,
-		worker:   worker,
-	}
-}
-
-func (fix *fixture) TearDown(c *gc.C) {
-	CheckStop(c, fix.worker)
-}
-
-func (fix *fixture) Guard(c *gc.C) (out fortress.Guard) {
-	err := fix.manifold.Output(fix.worker, &out)
-	c.Assert(err, jc.ErrorIsNil)
-	return out
-}
-
-func (fix *fixture) Guest(c *gc.C) (out fortress.Guest) {
-	err := fix.manifold.Output(fix.worker, &out)
-	c.Assert(err, jc.ErrorIsNil)
-	return out
-}
-
-func AssertUnlocked(c *gc.C, guest fortress.Guest) {
-	visited := make(chan error)
-	go func() {
-		visited <- guest.Visit(badVisit, nil)
-	}()
-
-	select {
-	case err := <-visited:
-		c.Assert(err, gc.ErrorMatches, "bad!")
-	case <-time.After(coretesting.LongWait):
-		c.Fatalf("abort never handled")
-	}
-}
-
-func AssertLocked(c *gc.C, guest fortress.Guest) {
-	visited := make(chan error)
-	abort := make(chan struct{})
-	go func() {
-		visited <- guest.Visit(badVisit, abort)
-	}()
-
-	// NOTE(fwereade): this isn't about interacting with a timer; it's about
-	// making sure other goroutines have had ample opportunity to do stuff.
-	delay := time.After(coretesting.ShortWait)
-	for {
-		select {
-		case <-delay:
-			delay = nil
-			close(abort)
-		case err := <-visited:
-			c.Assert(err, gc.Equals, fortress.ErrAborted)
-			return
-		case <-time.After(coretesting.LongWait):
-			c.Fatalf("timed out")
-		}
-	}
-}
-
-func CheckStop(c *gc.C, w worker.Worker) {
-	c.Check(worker.Stop(w), jc.ErrorIsNil)
-}
-
-func badVisit() error {
-	return errors.New("bad!")
 }
