+++ worker/fortress/manifold.go
@@ -14,11 +14,11 @@ import (
 //
 // Clients should access the fortress resource via Guard and/or Guest pointers.
 // Guest.Visit calls will block until a Guard.Unlock call is made; Guard.Lockdown
-// calls will block new Guest.Visits and wait until all current ones complete.
+// calls will block new Guest.Visits and wait until all active Visits complete.
 //
 // If multiple clients act as guards, the fortress' state at any time will be
 // determined by whichever guard last ran an operation; that is to say, it will
-// be impossible to reliably tell from outside, so please don't do that.
+// be impossible to reliably tell from outside. So please don't do that.
 func Manifold() dependency.Manifold {
 	return dependency.Manifold{
 		Start: func(_ dependency.GetResourceFunc) (worker.Worker, error) {
@@ -27,7 +27,7 @@ func Manifold() dependency.Manifold {
 		Output: func(in worker.Worker, out interface{}) error {
 			inFortress, _ := in.(*fortress)
 			if inFortress == nil {
-				return errors.Errorf("expected %T; got %T", inFortress, in)
+				return errors.Errorf("in should be %T; is %T", inFortress, in)
 			}
 			switch outPointer := out.(type) {
 			case *Guard:
